Remove ``move.*'' files because vick uses better function pointers now
The initial reason for having these files was to make it so that vick
could use raw function pointers to ``void (*)()'' methods.  This
worked pretty well but discouraged unit testing.  I switched to using

    std::function < void (contents&, boost::optional<int>) >

as it allowed for me to use prefix arguments and abstract over
``get_contents()'', vastly easing unit testing.

This commit also fixed the ``newmove.*'' files with the new way of using function pointers.